clean up deposition using ParallelFor with compile time optimizations
