+++ src/particles/deposition/PlasmaDepositCurrent.cpp
@@ -7,48 +7,16 @@
  */
 #include "PlasmaDepositCurrent.H"
 
+#include "particles/particles_utils/ShapeFactors.H"
+#include "particles/particles_utils/FieldGather.H"
 #include "particles/plasma/PlasmaParticleContainer.H"
-#include "PlasmaDepositCurrentInner.H"
+#include "particles/sorting/TileSort.H"
 #include "fields/Fields.H"
 #include "utils/Constants.H"
 #include "Hipace.H"
 #include "utils/HipaceProfilerWrapper.H"
-
-template<class...Args>
-void DepositCurrent_middle (bool outer_depos_loop, int depos_order_xy, bool use_laser,
-                            bool do_tiling, bool can_ionize, Args&&...args)
-{
-    if (outer_depos_loop && !use_laser && !do_tiling && !can_ionize) {
-        switch (depos_order_xy) {
-            case 0: return doDepositionShapeN<true, 0, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-            case 1: return doDepositionShapeN<true, 1, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-            case 2: return doDepositionShapeN<true, 2, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-            case 3: return doDepositionShapeN<true, 3, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-        }
-    } else if (!outer_depos_loop && !use_laser && !do_tiling && !can_ionize) {
-        switch (depos_order_xy) {
-            case 0: return doDepositionShapeN<false, 0, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-            case 1: return doDepositionShapeN<false, 1, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-            case 2: return doDepositionShapeN<false, 2, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-            case 3: return doDepositionShapeN<false, 3, false, false, false>(use_laser, do_tiling, can_ionize, args...);
-        }
-    } else if (outer_depos_loop) {
-        switch (depos_order_xy) {
-            case 0: return doDepositionShapeN<true, 0, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-            case 1: return doDepositionShapeN<true, 1, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-            case 2: return doDepositionShapeN<true, 2, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-            case 3: return doDepositionShapeN<true, 3, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-        }
-    } else {
-        switch (depos_order_xy) {
-            case 0: return doDepositionShapeN<false, 0, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-            case 1: return doDepositionShapeN<false, 1, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-            case 2: return doDepositionShapeN<false, 2, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-            case 3: return doDepositionShapeN<false, 3, true, true, true>(use_laser, do_tiling, can_ionize, args...);
-        }
-    }
-    amrex::Abort("unknow depos_order_xy: " + std::to_string(depos_order_xy));
-}
+#include "utils/Constants.H"
+#include "utils/GPUUtil.H"
 
 
 void
@@ -59,6 +27,7 @@ DepositCurrent (PlasmaParticleContainer& plasma, Fields & fields, const Laser& l
                 const PlasmaBins& bins, int bin_size)
 {
     HIPACE_PROFILE("DepositCurrent_PlasmaParticleContainer()");
+    using namespace amrex::literals;
 
     AMREX_ALWAYS_ASSERT_WITH_MESSAGE(
     which_slice == WhichSlice::This || which_slice == WhichSlice::Next ||
@@ -74,9 +43,8 @@ DepositCurrent (PlasmaParticleContainer& plasma, Fields & fields, const Laser& l
     amrex::Real const * AMREX_RESTRICT dx = gm.CellSize();
 
     const amrex::Real max_qsa_weighting_factor = plasma.m_max_qsa_weighting_factor;
-    const amrex::Real q = (which_slice == WhichSlice::RhoIons) ? -plasma.m_charge : plasma.m_charge;
+    const amrex::Real charge = (which_slice == WhichSlice::RhoIons) ? -plasma.m_charge : plasma.m_charge;
     const amrex::Real mass = plasma.m_mass;
-    const bool can_ionize = plasma.m_can_ionize;
 
     // Loop over particle boxes
     for (PlasmaParticleIterator pti(plasma, lev); pti.isValid(); ++pti)
@@ -94,19 +62,265 @@ DepositCurrent (PlasmaParticleContainer& plasma, Fields & fields, const Laser& l
         amrex::Vector<amrex::FArrayBox>& tmp_dens = fields.getTmpDensities();
 
         // extract the laser Fields
-        const bool use_laser = laser.m_use_laser;
         const amrex::MultiFab& a_mf = laser.getSlices(WhichLaserSlice::n00j00);
 
         // Offset for converting positions to indexes
         const amrex::Real x_pos_offset = GetPosOffset(0, gm, isl_fab.box());
         const amrex::Real y_pos_offset = GetPosOffset(1, gm, isl_fab.box());
-        const amrex::Real z_pos_offset = GetPosOffset(2, gm, isl_fab.box());
 
-        DepositCurrent_middle(Hipace::m_outer_depos_loop, Hipace::m_depos_order_xy,
-                              use_laser, Hipace::m_do_tiling, can_ionize,
-                              pti, isl_fab, jx_cmp, jy_cmp, jz_cmp, rho_cmp, chi_cmp,
-                              a_mf, tmp_dens, dx, x_pos_offset,
-                              y_pos_offset, z_pos_offset, q, mass, temp_slice,
-                              max_qsa_weighting_factor, bins, bin_size);
+        // Extract particle properties
+        auto& aos = pti.GetArrayOfStructs(); // For positions
+        const auto& pos_structs = aos.begin();
+        auto& soa = pti.GetStructOfArrays(); // For momenta and weights
+
+        amrex::Real * const AMREX_RESTRICT wp = soa.GetRealData(PlasmaIdx::w).data();
+        const amrex::Real * const AMREX_RESTRICT uxp =
+            soa.GetRealData(temp_slice ? PlasmaIdx::ux_temp : PlasmaIdx::ux).data();
+        const amrex::Real * const AMREX_RESTRICT uyp =
+            soa.GetRealData(temp_slice ? PlasmaIdx::uy_temp : PlasmaIdx::uy).data();
+        const amrex::Real * const AMREX_RESTRICT psip =
+            soa.GetRealData(temp_slice ? PlasmaIdx::psi_temp : PlasmaIdx::psi).data();
+
+        int const * const AMREX_RESTRICT a_ion_lev =
+            plasma.m_can_ionize ? soa.GetIntData(PlasmaIdx::ion_lev).data() : nullptr;
+
+        // Extract laser array from MultiFab
+        const Array3<const amrex::Real> a_laser_arr =
+            laser.m_use_laser ? a_mf[pti].const_array() : amrex::Array4<const amrex::Real>();
+
+        // Extract box properties
+        const amrex::Real invvol = Hipace::m_normalized_units ? 1._rt : 1._rt/(dx[0]*dx[1]*dx[2]);
+        const amrex::Real dx_inv = 1._rt/dx[0];
+        const amrex::Real dy_inv = 1._rt/dx[1];
+
+        const PhysConst pc = get_phys_const();
+        const amrex::Real clightinv = 1.0_rt/pc.c;
+        const amrex::Real charge_invvol = charge * invvol;
+        const amrex::Real charge_mu0_mass_ratio = charge * pc.mu0 / mass;
+
+        int n_qsa_violation = 0;
+        amrex::Gpu::DeviceScalar<int> gpu_n_qsa_violation(n_qsa_violation);
+        int* const AMREX_RESTRICT p_n_qsa_violation = gpu_n_qsa_violation.dataPtr();
+
+#ifdef AMREX_USE_OMP
+#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
+        {
+        const int ithread = omp_get_thread_num();
+#else
+        const int ithread = 0;
+#endif
+        AMREX_ALWAYS_ASSERT_WITH_MESSAGE(isl_fab.box().ixType().cellCentered(),
+            "jx, jy, jz, and rho must be nodal in all directions.");
+
+        const bool do_tiling = Hipace::m_do_tiling;
+
+        Array3<amrex::Real> const isl_arr =
+            do_tiling ? tmp_dens[ithread].array() : isl_fab.array();
+        const int jx  = (do_tiling && jx_cmp  != -1) ? 0 : jx_cmp;
+        const int jy  = (do_tiling && jy_cmp  != -1) ? 1 : jy_cmp;
+        const int jz  = (do_tiling && jz_cmp  != -1) ? 2 : jz_cmp;
+        const int rho = (do_tiling && rho_cmp != -1) ? 3 : rho_cmp;
+        const int chi = (do_tiling && chi_cmp != -1) ? 4 : chi_cmp;
+
+        int ntiley = 0;
+        if (do_tiling) {
+            const int ng = Fields::m_slices_nguards[0];
+            const int ncellx = isl_fab.box().bigEnd(0)-isl_fab.box().smallEnd(0)+1-2*ng;
+            const int ncelly = isl_fab.box().bigEnd(1)-isl_fab.box().smallEnd(1)+1-2*ng;
+            AMREX_ALWAYS_ASSERT(ncellx % bin_size == 0);
+            AMREX_ALWAYS_ASSERT(ncelly % bin_size == 0);
+            ntiley = ncelly / bin_size;
+        }
+
+        const int ntiles = do_tiling ? bins.numBins() : 1;
+#ifdef AMREX_USE_OMP
+#pragma omp for
+#endif
+        for (int a_itile=0; a_itile<ntiles; a_itile++){
+
+#ifndef AMREX_USE_GPU
+            if (do_tiling) tmp_dens[ithread].setVal(0.);
+#endif
+            // Get the x and y indices of current tile from its linearized index itile = itiley + itilex * ntiley
+            const int itilex = do_tiling ? a_itile / ntiley : 0;
+            const int itiley = do_tiling ? a_itile % ntiley : 0;
+            PlasmaBins::index_type const * const a_indices = do_tiling ? bins.permutationPtr() : nullptr;
+            PlasmaBins::index_type const * const a_offsets = do_tiling ? bins.offsetsPtr() : nullptr;
+            const int a_itilex_bs = do_tiling ? itilex * bin_size : 0;
+            const int a_itiley_bs = do_tiling ? itiley * bin_size : 0;
+
+            int num_particles = do_tiling ? a_offsets[a_itile+1]-a_offsets[a_itile]
+                                          : pti.numParticles();
+
+            if (Hipace::m_outer_depos_loop) {
+                num_particles *= (Hipace::m_depos_order_xy + 1);
+            }
+
+            // Loop over particles and deposit into jx_fab, jy_fab, jz_fab, and rho_fab
+            amrex::ParallelFor(
+                amrex::TypeList<
+                    amrex::CompileTimeOptions<0, 1, 2, 3>,  // depos_order
+                    amrex::CompileTimeOptions<false, true>, // outer_depos_loop
+                    amrex::CompileTimeOptions<false, true>, // can_ionize
+#ifdef AMREX_USE_GPU
+                    amrex::CompileTimeOptions<false>,       // do_tiling (disabled on GPU)
+#else
+                    amrex::CompileTimeOptions<false, true>, // do_tiling
+#endif
+                    amrex::CompileTimeOptions<false, true>  // use_laser
+                >{},  
+                {
+                    Hipace::m_depos_order_xy,
+                    Hipace::m_outer_depos_loop,
+                    plasma.m_can_ionize,
+                    do_tiling,
+                    laser.m_use_laser
+                },
+                num_particles,
+                [=] AMREX_GPU_DEVICE (int idx, auto depos_order, auto outer_depos_loop,
+                    auto can_ionize, auto c_do_tiling, auto use_laser) noexcept {
+                constexpr int depos_order_xy = depos_order.value;
+                // Using 1 thread per particle and per deposited cell is only done in the fast (x) direction.
+                // This can also be applied in the y direction, but so far does not show significant gain.
+                constexpr bool outer_depos_loop_x = outer_depos_loop.value;
+                constexpr int outer_depos_order_x_1 = outer_depos_loop_x ? (depos_order_xy + 1) : 1;
+                constexpr int inner_depos_order_x = outer_depos_loop_x ? 0 : depos_order_xy;
+
+                int ip = idx / outer_depos_order_x_1;
+
+                [[maybe_unused]] auto indices = a_indices;
+                [[maybe_unused]] auto offsets = a_offsets;
+                [[maybe_unused]] auto itile = a_itile;
+                if constexpr (c_do_tiling.value) {
+                    ip = indices[offsets[itile]+ip];
+                }
+
+                const int ox = idx % outer_depos_order_x_1;
+
+                const auto positions = pos_structs[ip];
+                if (positions.id() < 0) return;
+                const amrex::Real psi = psip[ip];
+                const amrex::Real xp = positions.pos(0);
+                const amrex::Real yp = positions.pos(1);
+                const amrex::Real vx_c = uxp[ip] / psi;
+                const amrex::Real vy_c = uyp[ip] / psi;
+
+                // calculate charge of the plasma particles
+                amrex::Real q_invvol = charge_invvol;
+                amrex::Real q_mu0_mass_ratio = charge_mu0_mass_ratio;
+                [[maybe_unused]] auto ion_lev = a_ion_lev;
+                if constexpr (can_ionize.value) {
+                    q_invvol *= ion_lev[ip];
+                    q_mu0_mass_ratio *= ion_lev[ip];
+                }
+
+                amrex::Real Aabssqp = 0._rt;
+                [[maybe_unused]] auto laser_arr = a_laser_arr;
+                if constexpr (use_laser.value) {
+                    doLaserGatherShapeN<depos_order_xy>(xp, yp, Aabssqp, laser_arr,
+                                                        dx_inv, dy_inv, x_pos_offset, y_pos_offset);
+                }
+
+                // calculate gamma/psi for plasma particles
+                const amrex::Real gamma_psi = 0.5_rt * (
+                    (1._rt + 0.5_rt * Aabssqp) / (psi * psi) // TODO: fix units
+                    + vx_c * vx_c * clightinv * clightinv
+                    + vy_c * vy_c * clightinv * clightinv
+                    + 1._rt
+                );
+
+                if ((gamma_psi < 0.0_rt || gamma_psi > max_qsa_weighting_factor) && ox == 0)
+                {
+                    // This particle violates the QSA, discard it and do not deposit its current
+                    amrex::Gpu::Atomic::Add(p_n_qsa_violation, 1);
+                    wp[ip] = 0.0_rt;
+                    pos_structs[ip].id() = -std::abs(positions.id());
+                    return;
+                }
+
+                for (int iy=0; iy <= depos_order_xy; ++iy) {
+                    for (int ix=0; ix <= inner_depos_order_x; ++ix) {
+                        int tx = 0;
+                        if constexpr (outer_depos_loop_x) {
+                            tx = ox;
+                        } else {
+                            tx = ix;
+                        }
+                        // --- Compute shape factors
+                        // x direction
+                        const amrex::Real xmid = (xp - x_pos_offset) * dx_inv;
+                        auto [shape_x, cell_x] =
+                            compute_single_shape_factor<outer_depos_loop_x, depos_order_xy>(xmid, tx);
+
+                        // y direction
+                        const amrex::Real ymid = (yp - y_pos_offset) * dy_inv;
+                        auto [shape_y, cell_y] =
+                            compute_single_shape_factor<false, depos_order_xy>(ymid, iy);
+
+                        [[maybe_unused]] auto itilex_bs = a_itilex_bs;
+                        [[maybe_unused]] auto itiley_bs = a_itiley_bs;
+                        if constexpr (c_do_tiling.value) {
+                            cell_x -= itilex_bs;
+                            cell_y -= itiley_bs;
+                        }
+
+                        const amrex::Real charge_density = q_invvol * wp[ip] * shape_x * shape_y;
+                        // wqx, wqy wqz are particle current in each direction
+                        const amrex::Real wqx  = charge_density * vx_c;
+                        const amrex::Real wqy  = charge_density * vy_c;
+                        const amrex::Real wqz  = charge_density * (gamma_psi-1._rt) / clightinv;
+                        const amrex::Real wq   = charge_density * gamma_psi;
+                        const amrex::Real wchi = charge_density * q_mu0_mass_ratio / psi;
+
+                        // Deposit current into isl_arr
+                        if (jx != -1) { // deposit_jx_jy
+                            amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, jx), wqx);
+                            amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, jy), wqy);
+                        }
+                        if (jz != -1) { // deposit_jz
+                            amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, jz), wqz);
+                        }
+                        if (rho != -1) { // deposit_rho
+                            amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, rho), wq);
+                        }
+                        if (chi != -1) { // deposit_chi
+                            amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, chi), wchi);
+                        }
+                    }
+                }
+            });
+#ifndef AMREX_USE_GPU
+            if (do_tiling) {
+                // If tiling is on, the current was deposited (see above) in temporary tile arrays.
+                // Now, we atomic add from these temporary arrays to the main arrays
+                amrex::Box srcbx = {{0, 0, 0}, {bin_size-1, bin_size-1, 0}};
+                amrex::Box dstbx = {{itilex*bin_size, itiley*bin_size, pti.tilebox().smallEnd(2)},
+                                    {(itilex+1)*bin_size-1, (itiley+1)*bin_size-1, pti.tilebox().smallEnd(2)}};
+                srcbx.grow(Fields::m_slices_nguards);
+                dstbx.grow(Fields::m_slices_nguards);
+                if (jx_cmp != -1) {
+                    isl_fab.atomicAdd(tmp_dens[ithread], srcbx, dstbx, 0, jx_cmp, 1);
+                    isl_fab.atomicAdd(tmp_dens[ithread], srcbx, dstbx, 1, jy_cmp, 1);
+                }
+                if (jz_cmp != -1) {
+                    isl_fab.atomicAdd(tmp_dens[ithread], srcbx, dstbx, 2, jz_cmp, 1);
+                }
+                if (rho_cmp != -1) {
+                    isl_fab.atomicAdd(tmp_dens[ithread], srcbx, dstbx, 3, rho_cmp, 1);
+                }
+                if (chi_cmp != -1) {
+                    isl_fab.atomicAdd(tmp_dens[ithread], srcbx, dstbx, 4, chi_cmp, 1);
+                }
+            }
+#endif
+        }
+#ifdef AMREX_USE_OMP
+        }
+#endif
+
+        n_qsa_violation = gpu_n_qsa_violation.dataValue();
+        if (n_qsa_violation > 0 && (Hipace::m_verbose >= 3))
+            amrex::Print()<< "number of QSA violating particles on this slice: " \
+                        << n_qsa_violation << "\n";
     }
 }
