+++ src/particles/pusher/PlasmaParticleAdvance.cpp
@@ -63,15 +63,15 @@ AdvancePlasmaParticles (PlasmaParticleContainer& plasma, const Fields & fields,
         const amrex::MultiFab& a_mf = laser.getSlices(WhichLaserSlice::n00j00);
 
         // Extract field array from MultiFab
-        amrex::Array4<const amrex::Real> const& a_arr = use_laser ?
-                                    a_mf[pti].array() : amrex::Array4<const amrex::Real>();
+        Array3<const amrex::Real> const& a_arr = use_laser ?
+            a_mf[pti].const_array() : amrex::Array4<const amrex::Real>();
 
-        const amrex::GpuArray<amrex::Real, 3> dx_arr = {dx[0], dx[1], dx[2]};
+        const amrex::Real dx_inv = 1._rt/dx[0];
+        const amrex::Real dy_inv = 1._rt/dx[1];
 
         // Offset for converting positions to indexes
         amrex::Real const x_pos_offset = GetPosOffset(0, gm, slice_fab.box());
         const amrex::Real y_pos_offset = GetPosOffset(1, gm, slice_fab.box());
-        amrex::Real const z_pos_offset = GetPosOffset(2, gm, slice_fab.box());
 
         auto& soa = pti.GetStructOfArrays(); // For momenta and weights
 
@@ -117,7 +117,6 @@ AdvancePlasmaParticles (PlasmaParticleContainer& plasma, const Fields & fields,
         amrex::Real * const Fpsi5 = soa.GetRealData(PlasmaIdx::Fpsi5).data();
         int * const ion_lev = soa.GetIntData(PlasmaIdx::ion_lev).data();
 
-        const int depos_order_xy = Hipace::m_depos_order_xy;
         const amrex::Real clightsq = 1.0_rt/(phys_const.c*phys_const.c);
 
         using PTileType = PlasmaParticleContainer::ParticleTileType;
@@ -145,8 +144,10 @@ AdvancePlasmaParticles (PlasmaParticleContainer& plasma, const Fields & fields,
             int const num_particles =
                 do_tiling ? offsets[itile+1]-offsets[itile] : pti.numParticles();
             amrex::ParallelFor(
+                amrex::TypeList<amrex::CompileTimeOptions<0, 1, 2, 3>>{},
+                {Hipace::m_depos_order_xy},
                 num_particles,
-                [=] AMREX_GPU_DEVICE (long idx) {
+                [=] AMREX_GPU_DEVICE (long idx, auto depos_order) {
                     const int ip = do_tiling ? indices[offsets[itile]+idx] : idx;
                     amrex::ParticleReal xp, yp, zp;
                     int pid;
@@ -162,16 +163,13 @@ AdvancePlasmaParticles (PlasmaParticleContainer& plasma, const Fields & fields,
                     if (do_update)
                     {
                         // field gather for a single particle
-                        doGatherShapeN(xp, yp,
-                                       ExmByp, EypBxp, Ezp, Bxp, Byp, Bzp, slice_arr,
-                                       exmby_comp, eypbx_comp, ez_comp, bx_comp, by_comp, bz_comp,
-                                       dx_arr, x_pos_offset, y_pos_offset, depos_order_xy);
+                        doGatherShapeN<depos_order.value>(xp, yp, ExmByp, EypBxp, Ezp, Bxp, Byp,
+                            Bzp, slice_arr, exmby_comp, eypbx_comp, ez_comp, bx_comp, by_comp,
+                            bz_comp, dx_inv, dy_inv, x_pos_offset, y_pos_offset);
 
                         if (use_laser) {
-                            doLaserGatherShapeN(xp, yp, 0._rt /* zp not used */,
-                                                Aabssqp, AabssqDxp, AabssqDyp, a_arr,
-                                                dx_arr, x_pos_offset, y_pos_offset, z_pos_offset,
-                                                depos_order_xy, 0);
+                            doLaserGatherShapeN<depos_order.value>(xp, yp, Aabssqp, AabssqDxp,
+                                AabssqDyp, a_arr, dx_inv, dy_inv, x_pos_offset, y_pos_offset);
                         }
                         // update force terms for a single particle
                         const amrex::Real q = can_ionize ? ion_lev[ip] * charge : charge;
@@ -194,8 +192,7 @@ AdvancePlasmaParticles (PlasmaParticleContainer& plasma, const Fields & fields,
                                            dz, temp_slice, ip, SetPosition, enforceBC );
                     }
                     return;
-                }
-                );
+                });
         }
     }
 }
